add PK between 2 id

+++ src/main/java/com/platzi/market/persistence/entity/Cliente.java
@@ -21,7 +21,7 @@ public class Cliente {
 
     private String apellidos;
 
-    private Double celular;
+    private Long celular;
 
     private String direccion;
